Replace dropdown import
