+++ packages/insomnia/src/ui/components/base/dropdown/dropdown.tsx
@@ -1,17 +1,14 @@
 import { PressResponder } from '@react-aria/interactions';
-import { Item, Section } from '@react-stately/collections';
-import { AriaButtonProps } from '@react-types/button';
 import type { AriaMenuProps, MenuTriggerProps } from '@react-types/menu';
 import type { Placement } from '@react-types/overlays';
 import classNames from 'classnames';
 import React, { CSSProperties, forwardRef, PropsWithChildren, ReactNode, useRef } from 'react';
 import { useMenuTrigger } from 'react-aria';
-import { mergeProps, useButton, useFocusRing } from 'react-aria';
 import { MenuTriggerState, useMenuTriggerState } from 'react-stately';
 import styled from 'styled-components';
 
 import { PlatformKeyCombinations } from '../../../../common/settings';
-import { Button as ThemedButton, ButtonProps } from '../../themed-button';
+import { DropdownButton } from './dropdown-button';
 import { DropdownHint } from './dropdown-hint';
 import { Menu } from './menu';
 import { Popover } from './popover';
@@ -135,92 +132,4 @@ const ItemContent = ({ icon, label, hint, className, onClick, children, iconStyl
   </StyledItemContainer>
 );
 
-interface StyledThemedButtonProps extends ButtonProps {
-  removePaddings?: boolean;
-  removeBorderRadius?: boolean;
-  disableHoverBehavior?: boolean;
-}
-
-const StyledThemedButton = styled(ThemedButton).attrs((props: StyledThemedButtonProps) => ({
-  variant: props.variant || 'text',
-  size: props.size || 'small',
-  radius: props.removeBorderRadius ? 0 : props.radius || '3px',
-}))`
-  height: 100%;
-  display: flex !important;
-  justify-content: space-between;
-  align-items: center;
-  padding: ${({ removePaddings = true }) =>
-    removePaddings ? 0 : 'var(--padding-xs) var(--padding-sm)'
-};
-
-  &:focus,
-  &:hover {
-    &:not(:disabled) {
-      ${({ variant = 'text', bg, disableHoverBehavior = true }) => {
-    if (!disableHoverBehavior) {
-      return;
-    }
-
-    if (variant === 'contained') {
-      return 'box-shadow: unset';
-    }
-
-    if (bg === 'default') {
-      return 'background-color: unset';
-    }
-
-    return 'background-color: unset';
-  }};
-    }
-  }
-
-  &:active:not(:disabled) {
-    ${({ variant, bg, disableHoverBehavior }) => {
-    if (!disableHoverBehavior) {
-      return;
-    }
-
-    if (variant === 'contained') {
-      return 'box-shadow: unset';
-    }
-
-    if (bg === 'default') {
-      return 'background-color: unset';
-    }
-
-    return 'background-color: unset';
-  }}
-  }
-`;
-
-type DropdownButtonProps = {
-  className?: string;
-  isPressed?: boolean;
-  style?: CSSProperties;
-  onClick?: () => void;
-  isDisabled?: boolean;
-} & ButtonProps & AriaButtonProps & StyledThemedButtonProps;
-
-const DropdownButton = forwardRef<{}, DropdownButtonProps>((props: DropdownButtonProps, ref: any) => {
-  const buttonRef = useRef(ref);
-  const { buttonProps, isPressed } = useButton({ ...props, onPress: props.onClick }, buttonRef);
-  const { focusProps } = useFocusRing();
-
-  return (
-    <StyledThemedButton
-      ref={buttonRef}
-      style={props.style}
-      isPressed={isPressed || props.isPressed}
-      variant={props.variant || 'text'}
-      size={props.size || 'small'}
-      {...mergeProps(buttonProps, focusProps, props)}
-    >
-      {props.children}
-    </StyledThemedButton>
-  );
-});
-
-DropdownButton.displayName = 'DropdownButton';
-
-export { Dropdown, Item as DropdownItem, Section as DropdownSection, ItemContent, DropdownButton };
+export { Dropdown, ItemContent };
