+++ packages/insomnia/src/ui/components/dropdowns/method-dropdown.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef, useCallback, useState } from 'react';
 import styled from 'styled-components';
 
 import { HTTP_METHODS, METHOD_GRPC } from '../../../common/constants';
-import { type DropdownHandle, Dropdown, DropdownButton, DropdownItem, DropdownSection, ItemContent } from '../base/dropdown/dropdown';
+import { type DropdownHandle, Dropdown, DropdownButton, DropdownItem, DropdownSection, ItemContent } from '../base/dropdown';
 import { showPrompt } from '../modals/index';
 
 const LOCALSTORAGE_KEY = 'insomnia.httpMethods';
