+++ packages/insomnia/src/ui/components/dropdowns/websocket-preview-mode.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 
 import { CONTENT_TYPE_JSON, CONTENT_TYPE_PLAINTEXT } from '../../../common/constants';
-import { Dropdown, DropdownButton, DropdownItem, ItemContent } from '../base/dropdown/dropdown';
+import { Dropdown, DropdownButton, DropdownItem, ItemContent } from '../base/dropdown';
 
 interface Props {
   previewMode: string;
